src/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.6
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
numpy==1.23.1
openpyxl==3.0.10
psycopg2-binary==2.9.3
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
urllib3==1.26.10
XlsxWriter==3.0.3
djangorestframework==3.13.1
pandas==1.4.3
